Scheduler: ignore discrepancy on number of meetings
The scheduler may start from a schedule that contains a different number of
meetings for a group than the requested number of slots. The scheduler drops
meetings beyond the requested number of slots but does not attempt to schedule
additional meetings if there aren't enough. It used to throw when that happened
but no longer does.