Calendar: always show joining information to holders of a W3C account
This update follows from discussions with Jean-Gui, Ian and Alex that stem from
confusion in the calendar UI around the term "event" which some interpreted to
mean "TPAC event", while the calendar system means "this calendar entry event".

The idea is now to show joining information to everyone registered to TPAC,
regardless of whether they're invited to this particular calendar entry.